Get rid of unused code
